+++ shared/component/tab/editor/child/attachments/src/commonTest/kotlin/com/sedsoftware/tackle/editor/attachments/stubs/EditorAttachmentsApiStub.kt
@@ -26,7 +26,7 @@ class EditorAttachmentsApiStub : StubWithException(), EditorAttachmentsGateways.
 
     override suspend fun sendFile(
         file: PlatformFileWrapper,
-        onUpload: (Int) -> Unit,
+        onUpload: (Float) -> Unit,
         thumbnail: PlatformFileWrapper?,
         description: String?,
         focus: String?,
