+++ shared/component/tab/editor/src/commonMain/kotlin/com/sedsoftware/tackle/editor/EditorTabComponentGateways.kt
@@ -14,7 +14,7 @@ interface EditorTabComponentGateways {
 
         suspend fun sendFile(
             file: PlatformFileWrapper,
-            onUpload: (Int) -> Unit = {},
+            onUpload: (Float) -> Unit = {},
             thumbnail: PlatformFileWrapper? = null,
             description: String? = null,
             focus: String? = null,
