+++ shared/component/tab/editor/src/commonMain/kotlin/com/sedsoftware/tackle/editor/integration/attachments/EditorAttachmentsComponentApi.kt
@@ -11,7 +11,7 @@ internal class EditorAttachmentsComponentApi(
 
     override suspend fun sendFile(
         file: PlatformFileWrapper,
-        onUpload: (Int) -> Unit,
+        onUpload: (Float) -> Unit,
         thumbnail: PlatformFileWrapper?,
         description: String?,
         focus: String?,
